Modularize tournament match form and result display into reusable components
